\name{get.abMat}
\alias{get.abMat}
\title{
   Default set of parameters to pass to \code{\link{Estim_Simulation}}
}
\description{
  Default set of parameters to pass to \code{\link{Estim_Simulation}},
  inspired by the one used by Koutrevelis (1980) in his simulation
  procedure.
}
\usage{
get.abMat()
}
\value{
  a 2-columns matrix containing a wide range of parameters \eqn{\alpha}
  and \eqn{\beta} covering the entire parameters space.
}
\keyword{data-functions}
