\name{getTime_}
\alias{getTime_}
\title{
  Read time
}
\description{
  Reads the time when the function is called.
}
\usage{
getTime_()
}

\value{
  a \code{numeric}.
}
\seealso{
  \code{\link{PrintDuration}},
  \code{\link{PrintEstimatedRemainingTime}},
  \code{\link{ComputeDuration}}
}
\examples{
ti <- getTime_()
}
\keyword{general-functions}
