% see Repeated descriptions of re-exported functions #408 at
%   https://github.com/klutometis/roxygen/issues/408
\docType{import}
\name{StableEstim_reexports}
\alias{StableEstim_reexports}
\alias{rstable}

\title{Objects exported from other packages}
\description{
  These objects are imported from other packages. Follow the links
  below to see their documentation.

  \describe{
    \item{rstable}{\code{\link[stabledist]{rstable}}}
  }
}
\keyword{internal}
